LayoutTests/fast/text/emoji-num-glyphs.html

<!DOCTYPE html>
<html>
<head>
<style>
#console {
	font: 16px Times;
}
</style>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body style="font: 100px Times;">
<span id="control">&#x26F9;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F466;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F467;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F467;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F468;&zwj;&#x1F467;&zwj;&#x1F467;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F466;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F467;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F467;&zwj;&#x1F466;</span>
<span class="testElement">&#x1F469;&zwj;&#x1F467;&zwj;&#x1F467;</span>



<span class="testElement">&#x1F3F3;&#xFE0F;&zwj;&#x1F308;</span>
<span class="testElement">&#x1F3F3;&zwj;&#x1F308;</span>



<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FB;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FC;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FD;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FE;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;&#x1F3FF;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C3;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C3;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FB;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FC;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FD;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FE;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;&#x1F3FF;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3C4;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3C4;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FB;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FC;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FD;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FE;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;&#x1F3FF;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CA;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CA;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FB;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FC;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FD;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FE;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;&#x1F3FF;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CB;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F3CC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F3CC;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FB;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FC;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FD;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FE;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;&#x1F3FF;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2640;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F46E;&zwj;&#x2642;</span>
<span class="testElement">&#x1F46E;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FB;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FC;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FD;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FE;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;&#x1F3FF;</span>
<span class="testElement">&#x1F471;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F471;&zwj;&#x2640;</span>
<span class="testElement">&#x1F471;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F471;&zwj;&#x2642;</span>
<span class="testElement">&#x1F471;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FB;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FC;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FD;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FE;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;&#x1F3FF;</span>
<span class="testElement">&#x1F473;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F473;&zwj;&#x2640;</span>
<span class="testElement">&#x1F473;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F473;&zwj;&#x2642;</span>
<span class="testElement">&#x1F473;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FB;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FC;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FD;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FE;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;&#x1F3FF;</span>
<span class="testElement">&#x1F477;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F477;&zwj;&#x2640;</span>
<span class="testElement">&#x1F477;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F477;&zwj;&#x2642;</span>
<span class="testElement">&#x1F477;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FB;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FC;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FD;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FE;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;&#x1F3FF;</span>
<span class="testElement">&#x1F481;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F481;&zwj;&#x2640;</span>
<span class="testElement">&#x1F481;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F481;&zwj;&#x2642;</span>
<span class="testElement">&#x1F481;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FB;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FC;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FD;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FE;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;&#x1F3FF;</span>
<span class="testElement">&#x1F482;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F482;&zwj;&#x2640;</span>
<span class="testElement">&#x1F482;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F482;&zwj;&#x2642;</span>
<span class="testElement">&#x1F482;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FB;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FC;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FD;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FE;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;&#x1F3FF;</span>
<span class="testElement">&#x1F486;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F486;&zwj;&#x2640;</span>
<span class="testElement">&#x1F486;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F486;&zwj;&#x2642;</span>
<span class="testElement">&#x1F486;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FB;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FC;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FD;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FE;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;&#x1F3FF;</span>
<span class="testElement">&#x1F487;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F487;&zwj;&#x2640;</span>
<span class="testElement">&#x1F487;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F487;&zwj;&#x2642;</span>
<span class="testElement">&#x1F487;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FB;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FC;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FD;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FE;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;&#x1F3FF;</span>
<span class="testElement">&#x1F575;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F575;&zwj;&#x2640;</span>
<span class="testElement">&#x1F575;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F575;&zwj;&#x2642;</span>
<span class="testElement">&#x1F575;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FB;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FC;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FD;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FE;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;&#x1F3FF;</span>
<span class="testElement">&#x1F645;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F645;&zwj;&#x2640;</span>
<span class="testElement">&#x1F645;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F645;&zwj;&#x2642;</span>
<span class="testElement">&#x1F645;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FB;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FC;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FD;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&#x1F3FE;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F646;&zwj;&#x2640;</span>
<span class="testElement">&#x1F646;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F646;&zwj;&#x2642;</span>
<span class="testElement">&#x1F646;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F647;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F647;&zwj;&#x2640;</span>
<span class="testElement">&#x1F647;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F647;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64B;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64D;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F64E;&#x1F3FF;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2640;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F64E;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FB;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FC;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FD;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FE;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;&#x1F3FF;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6A3;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6A3;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FB;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FC;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FD;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FE;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;&#x1F3FF;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B4;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B4;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FB;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FC;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FD;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FE;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;&#x1F3FF;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B5;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B5;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FB;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FC;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FD;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FE;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;&#x1F3FF;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2640;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x1F6B6;&zwj;&#x2642;</span>
<span class="testElement">&#x1F6B6;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FB;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FB;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FC;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FC;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FD;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FD;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FE;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FE;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&#x1F3FF;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;&#x1F3FF;</span>
<span class="testElement">&#x26F9;&zwj;&#x2640;&#xFE0F;</span>
<span class="testElement">&#x26F9;&zwj;&#x2640;</span>
<span class="testElement">&#x26F9;&zwj;&#x2642;&#xFE0F;</span>
<span class="testElement">&#x26F9;&zwj;&#x2642;</span>
<span class="testElement">&#x26F9;</span>
<script>
var controlWidth = document.getElementById("control").offsetWidth;
let elements = document.getElementsByClassName("testElement");
for (var i = 0; i < elements.length; ++i) {
	shouldBe("elements[i].offsetWidth", "controlWidth");
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
